requirements.txt
datasets>=2.17.1
fortepyan==0.4.2
hydra-core>=1.3.2
midi_tokenizers @ git+https://github.com/Nospoko/midi-tokenizers@v2.0
pandas~=2.1.4
python-dotenv==1.0.0
streamlit==1.37.0
streamlit-pianoroll>=0.7.1
torch>=2.1.2
transformers>=4.39.3
wandb>=0.15.0
